DRTII-1000 Finish generic feed arrivals actor
